requirements.txt
google-generativeai
python-dotenv
markdown
beautifulsoup4
